Don't break very long words (eg urls)
The greedy-wrap js library can't do this so it had to be replaced by word-wrap.
word-wrap has a bug if a newline character appears at the wrapping column. One
unit test had to be removed because of this, but it doesn't affect the rewrap's
output because we already strip the newline characters and replace them with
spaces beforehand in the code