Wrap affected blocks immediately
Instead of working out which of all blocks are affected by selections
and then passing the results to the wrapping module, it now works out if
the block is affected and calls the wrap function from the wrapping
module immediately, passing the result to the output lines.

This makes things simpler, and the next stage will be to look at each
section right after parsing instead of parsing everything first.

This code should maybe now be done in another file than Selections.fs
(Main.fs?)