LaTeX: replace inline commands list with block
Removes the list of "inline" commands, which was never going to be
complete anyway. Instead have any tag that's alone on a line keep the
line breaks before and after.

This works for most things (section, usepackage etc). For a small number
of commands where we want to keep the line break before even if there's
text after it on the same line, we have a list of "block" commands.
(Currently only \[, \item).)