Make autowrap work with CJK languages
Quite a few changes here. The methods to execute a standard wrap and
auto-wrap were quite similar, but they have diverged, each using only
the data relevant to them.

Auto-wrap now has some different checks. Instead of just space or enter,
it now checks for any break-before or break-after char (whitespace and
CJK characters) in the new text. We could have just triggered on any
character, but consideration had been taken for IMEs: it mustn't trigger
until input is finished and the roman characters have been transformed
into native characters.

DocState: DocState was a handy but overloaded record type. Now it's only
used in the ruler-switching logic. The language field has been removed
since it's not relevant here. The wrapping functions no longer use it as
an input. Instead, they use a new File record (couldn't think of a
better name) which is just the filename and language, used to select the
parser.

For auto-wrap: instead of calling cursorbeforeWrappingColumn and then
autoWrap, this has been merged into one function: maybeAutoWrap, which
checks the inputs given and then either returns an empty edit if no
autowrap was necessary, or a valid edit if it was. Instead of getting
the selection from DocState, it now takes selection-agnostic position
of input and new text parameters.

This means however that we need to calculate where the selection should
be after the edit (though not after the rewrap). Therefore the Edit
record now has a selections field, where this is passed back after the
autowrap is done. For a normal rewrap, the input selections are passed
straight back. Later we may put the diffing logic into core, so that the
selections passed back are exactly what they should be after the rewrap.

The client code has changed accordingly: both clients now just use two
separate functions for calling the standard wrap and auto-wrap. Since
the auto-wrap doesn't deal with ruler-switching, a new function
getWrappingColumn was added to core, which just gets either the current
column for the document or the first in the list of rulers given.
getDocWrappingColumn (which took a DocState and maybe changed the
column) has been renamed to maybeChangeWrappingColumn.

VSCode: since we no-longer need the separate Settings.getWrappingColumns
function, the settings module is now just a single function export.