VSC: Look up comment markers for unknown file types
In VSCode, for file types that aren't known to Rewrap, search through
extensions for a language configuration file that has comment marker
information that we can use to support it.

Reference: https://code.visualstudio.com/docs/extensionAPI/extension-points#_contributes.languages.
There isn't an API in VSCode to get this data for other languages
(vscode issue #2871) so we have to hack something together ourselves.

Not implemented yet in VS (not sure if it ever will be), just the
minimum code to make it compile. Have made CustomMarkers a nullable
class type to allow this.