Fully support VS2019 (resolves #122)
- Updates source.extension.vsixmanifest to include VS v16.0
- Migrates Package and Commands to use to AsyncPackage for loading the
  extension

In `AutoWrapCommand` we've had to change to program against
IVsWritableSettingsStore instead of WritableSettingsStore since we no
longer have a IServiceProvider package to pass to the
ShellSettingsManager() constructor. There is a
ShellSettingsManager(IVsSettingsManager) constructor that we could have
otherwise used, but this is only supported in VS2017+.

Was in PR #127.

Resources:
- https://devblogs.microsoft.com/visualstudio/how-to-upgrade-extensions-to-support-visual-studio-2019/
- https://github.com/Microsoft/VSSDK-Extensibility-Samples/tree/dfc288fb2b4465a45bc5e6f5f23d82ca35cd7cdb/AsyncPackageMigration