VS: Change supported versions to up to 17.0 (excl)
Starting with VS 2019, the minor version number of the product may now
be incremented. So now that VS 2019 has updated to v16.1, our supported
range of [14.0,16.0] no longer supports it.

Have changed the supported version ranges to [14.0,17.0) to support
everything up to (not including) v17. Open ended version ranges, eg:
`[14.0,)`, are unfortnumarely not allowed on the marketplace.

Reference:
https://devblogs.microsoft.com/visualstudio/visual-studio-extensions-and-version-ranges-demystified/