Add extra test script onto prepublish for exit code
The spec tests get run as part of the fable compile, but that doesn't
return a non-zero exit code if they fail. Adding an extra `npm test`
onto the end of the vscode:prepublish script so that CI will fail if the
tests fail. The tests will be run twice but they don't take <1s. They
need to stay part of the fable compile processess though for the
dev/watch mode script.

Also adding --silent to save the extra npm error lines