VS: Upgrade and fix dependencies (#195)
Upgrades dependencies FSharp.Core to 4.7.1 and System.ValueTuple to
4.5.0.

Migrates from packages.config to PackageReference for both projects.
This change however caused the F#Core and ValueTuple dll's to no longer
be built into the .vsix file. The solution was found here:
https://stackoverflow.com/questions/57449047/how-should-i-include-package-references-in-a-vsix-project.
F#Core could also be removed as a dependency of the VS project; it's now
just on the Core project (as would be expected) and the dll still gets
built into the .vsix file.