Modify "Rewrap at column..." command to make "unwrapping" easier (#107)
The aim was to add an "Unwrap" command (#107).

The existing code made this very easy: the existing `doWrap` function
can already take a custom column (for the "custom column..." command),
so just use this but make the column to wrap at a very large value.

I therefore realized that "unwrapping" could already be done with the
"custom column..." command, if the user entered a very large value.

But this wouldn't be great UX, However instead of adding another
relatively uncommonly used command, we can repurpose this one.

The command has been renamed to "Rewrap/Unwrap Text At Column...", and
now the dialog box opens up with a blank value, and if the user leaves
it blank then an "unwrap" will be done (this is explained with
additionaly text).

The box then passes this back as a wrapping column of < 1. Changes have
been made to the core code so that if whenever a wrap is done with
column < 1, it's instead done with column at effectively infinity
(Int32.MAX)