Stop caching settings for each document (fixes #99)
VSCode issue #48225 has been fixed, and retrieving settings is not
expensive, so it makes little sense to cache them for each document.

This and another piece of logic was also causing the issue where a
change to the wrapping column in settings wouldn't be picked up until
the editor was restarted (also in VS).

The other logic was that, while the current/last used wrapping column
for each document was being stored, we were just using that value when a
new wrap was requested, and not checking it against the new passed-in
list of available columns, to see if it was still valid (in
`getWrappingColumn`). Now, we do that check, and if it's not found, we
use the first value in that list instead.

However we do now (in VSCode) instead cache warnings for invalid values,
because otherwise, a warning would potentially be written to the console
for each keypress if a setting was invalid, flooding the console. VSCode
mitigates this by collapsing identical warnings, but it still wouldn't
be ideal.