VS: Fix an index out of bounds bug (#159)
Was occurring if no wrapping column and also no rulers were set. VS
Editor now adds a default value of 80 in this case (the same as happens
in VSC).

`getWrappingColumn` and `maybeChangeWrappingColumn` now expect rulers to
be a non-empty array. This feels maybe a little dangerous but the
number of places these are called is limited and it feels a little
overkill to convert rulers into a list and then nonempty list in these
functions.