Fix autowrap in embedded sections in html files (#162)
When an autowrap is done we only pass in lines up to & including where
the cursor is (an optimization). In the case of an embedded section in
an html file that might look like this:

```
<script>
    // Some long line we're writing that and autowrap should trigger
```

The code that normally looks at embedded <script> sections in an html
file assumes (in a normal case) that the first and last lines would both
just be (opening & closing) script tags, and that they both should be
ignored. This caused it to ignore the last line that in autowrap cases
might be the comment line we're writing instead.

This change checks the last line again for the presence/absence of the
ending script tag and parses accordingly. It's a bit of a hack until I
think of something cleaner.