Rename Extensions -> Prelude, move Nonempty
It was wrong for other modules to use `open Nonempty`. It was necessary
to bring the Nonempty type into scope but bringing in all the other
functions from that module was wrong. Since we can't pick what we want
to import, the Nonempty type is moved into Prelude, and `open Nonempty`
is no longer required at the top of modules, and the functions
themselves are already qualified throughout the code.

Also renames the `option` function to `maybe` and makes it available
directly from the Prelude