Support new v1.43 VSCode ruler objects
From v1.43 of VSCode the previous `editor.rulers` (for either general
or language-specific) has been changed to also accept the following
object which takes in a specific color for each ruler:
`{"column": 80, "color": "#ff4081"}`

More info on the PR for this change:
https://github.com/microsoft/vscode/pull/88453

The Rewrap extension only had support for the number rulers, so this
commit changes that. It takes care of figuring out if the ruler is
using the bare number or the new object, and if using the object, will
convert it down to just the column number to keep the rest of the code
as is.