Preserve linebreaks after lines ending in 2 spaces for all types
This was just for markdown, now extended to plaintext, latex, html/xml
and go.

Each type handles this independently. A future version may be able to
move this feature into shared code.

That this is done at the parsing and not at the wrapping stage means a
paragraph with a line-break in the middle is actually seen as two
paragraphs. This affects:
1. The scope of what is wrapped when a wrap is requested with an empty
   cursor.
2. Line indents. Currently paragraphs remember the indent of their first
   and second lines.

I think it's more obvious this way (especially point 1) but the other
way could be argued too.