Markdown blockquotes: fix disappearing space in prefix (#204)
See added test for simplified version. For a blockquote block, we take
the prefixes for the head & tail lines of the wrapped result from the
1st & 2nd lines of the original, just as in Everything else. However
here because of the "blank" 2nd line (just a ">"), it's taking that for
the rest of the blockquote section.

There's probably a better fix but not doing it in this version now. The
workaround is just to check if there's a ">" in the 2nd line. If not we
use that for the tail lines, otherwise we use the first line prefix.
This is what this code is really for.

This means eg this won't get the prefixes preserved correctly but it's
such a corner case:

    > first line
    >second line