Add Functor typeclass, streamline Prelude & Nonempty modules
Wanted to put everything related to the Nonempty list type in one file.
The type and a few functions would be available automatically
everywhere, with then a nested Nonempty module for the specific
functions.

However the typeclass stuff didn't work with the Nonempty type in a
separate file. The map function in Prelude couldn't find the Nonempty
type to match it. Not strictly an ordering issue because it can still
find the type, even though it's below the function in the Prelude file
(which isn't a rec module).