Markdown blockquotes: preserve prefixes per-line (#204)
The current approach of, for any wrappable block, determining separate
for only the head and tail lines, is failing us in the cases seen in
issue 204.

This modifies Block.splitUp function so that it takes a list of prefixes
(for each line of the original text) instead of just two. It then
matches them up with the sub-blocks that the inner parser creates.

Currently only the markdown blockquotes make use of this.

This approach of preserving prefixes per-line will cause some slight
differences in behavior but it may be needed for correctness and should
be tried in other types of blocks too.