Change rules for how comment content indent is determined (#203)
Changes the rules for indentation of comment content. Now, instead of
the indent being determined from the first line with text, the whole
block is taken as-is. Meaning the "base" indent is taken from the
last-indented line with text. This is necessary to allow comments to
start with an indented section, as seen in #203.

A knock-on change is that a line-comment block is now determined by line
comment markers that have the same indent, ie if a line has its comment
markers at a different column from the previous line, it's treated as
the start of a new block, rather than continuing the current as before.
This is however a change we would have to make anyway to support end of
line comments.

Another change is that for comments, the prefix of each line is
preserved (per content block), instead of being taken from the first
line with text. This a continuation of the work in d541fe8.

Code changes
------------

Move char/string functions from Wrapping.fs and Main.fs so that they can
be used earlier (during parsing). Put them in Line.fs since that has
other string-related functions. charWidth, strWidth, strWidth'.
charWidthEx is now just charWidth in Line.js; wrapping still uses the
same charWidth for now.

Line.fs:
- Use new style for existing functions
- Add new Line type

Prelude.fs:
- Add option workflow
- Add List.spanMaybes

Nonempty.fs:
- Change fromListUnsafe to fromSeqUnsafe
- Add mapLast, mapFold and spanMaybes functions

Parsing.Core.fs:
- Types/functions here have been made generic so they can work both on
  strings (old) and Lines (new). Not sure it's all necessary but these
  will be replaced eventually. This meant adding extra string type in a
  lot of other code.

Parsing.Comments.fs:
- Almost completely rewritten; however a bit messy since it's gradually
  bringing in new concepts (line-by-line parsing, the Line type), while
  not doing it completely that way yet.

Parsing.Documents.fs
- In eg javadoc-style comments, prefix chars (eg '*') are now always
  optional, to avoid problems if they're not added (eg in the PHP regex
  they were mandatory). To ensure this, the string is now just a list of
  allowed chars, which is added into a regex later, inside '[]'.