Fix issues that broke extension-contributed languages
Upgrading to Fable 2 broke this for two reasons.

An object of the `CustomMarkers` type must now be created using its
constructor; in the Fable-generated JS its not expected just to be a
POJO.

Also a bug in Fable with currying (2116) means that Parsing.Documents.fs
line 267 needed to be rewritten to avoid using `List.choose`.