Set as a UI extension (#231, #201)
Wasn't being detected as a UI extension by VSCode, so when working with
remote workspaces, it would have to be installed into the remote
workspace.

This also meant that other locally-installed language extensions weren't
being found (vscode.extensions.all didn't contain them?) causing
issue #201, which should be fixed by this.

The option for it to be installed as a workspace extension is still
available but not sure if it should be, and where this would be
necessary.