Move docs from wiki to /docs & use docsify
Have seen other repos do this and now doing it here. Having the docs in
the same repo helps to keep it at hand and it can be more easily kept in
sync when features are changed/added.

Using docsify allows more customizing in css, search etc.

Also moves logo files from /images to /docs images.

The whole /docs folder is put in the .vscodeignore whitelist. The logo
images in there is needed, but not sure yet if the docs should go in the
released extension or it's just bloat (especially since specs will
probably be moved there too).