Write core testrunner in F#
Adds a new F# project: Core.Test. This runs the spec tests against the
core F# code directly, allowing for easier debugging.

It will eventually also be transpiled to JS and replace the JS
testrunner file /tests.js, but that will come with the upgrade to Fable
compiler v3.

The new testrunner does some extra checking of the test sources, which
leads to some tests having to be fixed where the column marker in the
expected section is not on the same column as in the input.

It also counts 14 more tests than the old, because it for tests that
have a "normal" and reformat expected output, it counts that as 2 tests
where the old counts them as 1.