Change CustomMarkers 'null' value
Goal: The JS code currently needs to import the CustomMarkers
type/constructor from Types.js. Remove this need so that Main.js can be
the single point of entry, which will make it easier when bundling up
the core with Parcel etc.

Removing the AllowNullLiteral attribute from the CustomMarkers type and
making it a plain record, as well as removing the Option types from that
record, means it can be more easily created in C# code, and created as a
POJO in JS code.

The Option types on the line & block members were also redundant, since
an empty string is also an empty (invalid) value. Empty strings were not
checked-for in addCustomLanguage, leading to a runtime error if they
were supplied.

Now addCustomLanguage is changed to maybeAddCustomLanguage, and it only
adds the language if it has a valid line or block member.