New build system with Fable 3
- Upgrades to Fable 3
- Uses Parcel to build the vscode extension, creating an optimized
  bundle that reduces the size of the extension significantly
- Moves the package.json for the vscode extension from the root folder
  into the /vscode folder; that folder is now the "root" for that
  extension.
- Replaces NPM scripts for building/testing with a custom shellscript

Fable 3 is installed as a dotnet tool, replacing the
fable-splitter/compiler npm dependencies. Build steps are as follows:
- Fable compiles core F# files 1:1 into .js file, via the Core.Test
  project. This creates a runnable core test-runner that can be run
  with node, while also generating the core library for the vscode
  extension.
- For the vscode extension written in TypeScript, the TS is checked
  but no output files are written; Parcel is then used to bundle up
  everything into a single /bin/Extension.js file. This can be done in
  dev (no minification, with source maps) or production (minification,
  no source maps) modes.

The old js specs test-runner (tests.js) has therefore been replaced
with this Fable-compiled version. The file Native.js contains
IO-related FFI functions for javascript.

Because Fable 3 only compiles into esmodules, not commonjs, and
vscode still requires commonjs modules, we can't (yet) just compile the TS
and run the extension like that. So Parcel is always used in the vscode
build, even in development, to take care of everything for us.

package.json is now split. The package.json in the repository root is
now only for general dependencies for tooling. The package.json for the
vscode extension has been moved into /vscode; making that folder now the
root for the extension. Some required files (logo, license) had to have
copies made into this folder. Since the root of the /vscode folder
itself was now getting cluttered, the TypeScript source files have now
been moved into a /src subfolder. (This also makes it easier for the
import path in the TS source back to the core module, since the source
and output files are now at the same level)

NPM scripts aren't great for a complex build system. They have been
replaced (for the moment) by a shellscript `./do` that deals with
various build commands. Will look to move this to Fake.