Wrapping: fix where trailing spaces go over the wrapping column
Trailing spaces on the last line of any wrapping block should be
preserved. But when there was enough whitespace to wrap onto the next
line, this was creating a new empty line.

This change only wraps once a non-whitespace character is found, thus
trailing whitespace on the last line is always preserved, even if it
goes over the wrapping column.