Update and fix vscode tests
Stage 1: get tests to work again:
- Move tests from /vscode-test to /vscode/test
- Change dependency from 'vscode-test' to renamed & newer
  '@vscode/test-electron' (this fixes the error)
- Change vscode binary download location to .obj/vscode-test
- Use option `reuseMachineSettings`. Will have to see if this causes
  problems.

Stage 2: get a wrapping test working:

This means a change in how the extension works, because otherwise you
can't (easily) test the wrapping command, because after awaiting the
call to workspace.executeCommand("rewrap.rewrapComment"), the wrap edit
for the wrap won't have been done yet (because `rewrapComment` is a
TextEditor command).

For TextEditor commands, you're supposed to use the supplied
TextEditorEdit object, rather than call editor.edit() as we were. So
everything has been rewroked to do that. It may involve a slight
timing/performance improvement. It doesn't affect auto-wrap though since
that's not done in a TextEditor command.

It does though mean that another method is needed to invoke the
post-wrap selections fixing, and it's storing the data in a global
object and listening to onDidChangeTextDocument events. It's quite ugly.