VSC: Build edits from diff
Instead of just `editBuilder.replace`ing the old lines with new, uses
fast-diff to run a diff on them and create `editBuilder.insert` and
`editBuilder.delete` operations from that. This has the large benefit
the selections now move to the right places.

This also means we can calculate in advance where the post-wrap
selections will be, and call `saveDocState` immediately, instead of the
ugly method before of listening to document changed events. The only
thing is we 'guess' the document's version after wrap will be now+1.

A further improvement would be to refactor the code that deals with
adjusting selections into a separate class, and only calling it if
saveState is true (ie not for an auto-wrap).