Make comment marker regex patterns case-insensitive
Makes it easier for Windows batch files, where "REM" can be any case.
There aren't many languages that have words in the comment markers (Ruby
and Liquid being two). Not sure if in those cases case matters, but it's
probably no harm this way.

If this needs to be reversed, then using a regex `[Rr][Ee][Mm]` for
batch files isn't hard.