Fix error when using "Rewrap At... command" (#324)
Unlike the `rewrapCommentCommand` function, `rewrapCommentAtCommand` has
to be async, because it shows the input popup first. So it can't use the
editBuilder passed directly in (this gives the error "Edit is only valid
while callback runs").

So we have to go back to the old way of using the `editor.edit(...)`
function here. Don't think any harm should come from this.