VSC: Add "browser" field to package.json
This is to trick the Marketplace into thinking that the extension
supports VS Code on the Web. Currently the stable version doesn't, but
it's being tested on the pre-release version. However the Marketplace
always check the package.json of the stable version and then fails to
install, even if you tell it to install the pre-release version.

Also tells parcel to ignore this field (otherwise it complains it can't
build two bundle targets with the same target filename).